Corrected Updater URL

+++ src/CharacterCreator.java
@@ -2,11 +2,11 @@ import javax.swing.SwingUtilities;
 
 public class CharacterCreator {
 	
-	static double thisVersion = 1;
+	static int thisVersion = 1;
 
 	public static void main(String[] args) {
 		try {
-			if (Double.parseDouble(UpdateChecker.getLatestVersion()) > thisVersion) {
+			if (Integer.parseInt(UpdateChecker.getLatestVersion()) > thisVersion) {
 				new UpdateInfo(UpdateChecker.getWhatsNew());
 			} else {
 				// Initial Data List Setup work:
